% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{add_annotation}
\alias{add_annotation}
\title{Add annotation to layr visualization}
\usage{
add_annotation(layr_obj, annotation)
}
\arguments{
\item{layr_obj}{A layr object}

\item{annotation}{An annotation object}
}
\value{
Modified layr object
}
\description{
Add annotation to layr visualization
}
