% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aestetics.R
\name{aes_auto}
\alias{aes_auto}
\title{Automatic aesthetic mapping with intelligent defaults}
\usage{
aes_auto(data, x, y = NULL, ...)
}
\arguments{
\item{data}{The data frame}

\item{x}{X variable}

\item{y}{Y variable}

\item{...}{Additional mappings}
}
\value{
A layr aesthetic mapping object
}
\description{
Automatic aesthetic mapping with intelligent defaults
}
