% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aestetics.R
\name{aes_semiotic}
\alias{aes_semiotic}
\title{Create aesthetic mapping for layr}
\usage{
aes_semiotic(...)
}
\arguments{
\item{...}{Aesthetic mappings}
}
\value{
A layr aesthetic mapping object
}
\description{
Create aesthetic mapping for layr
}
