% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{annotation_highlight}
\alias{annotation_highlight}
\title{Create a highlight annotation}
\usage{
annotation_highlight(data, label = NULL, story = NULL, ...)
}
\arguments{
\item{data}{Data to highlight}

\item{label}{Label for highlight}

\item{story}{Narrative description}

\item{...}{Additional parameters}
}
\value{
An annotation object
}
\description{
Create a highlight annotation
}
