% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{annotation_trend_line}
\alias{annotation_trend_line}
\title{Create a trend line annotation}
\usage{
annotation_trend_line(method = "lm", story = NULL, ...)
}
\arguments{
\item{method}{Statistical method for trend}

\item{story}{Narrative description}

\item{...}{Additional parameters}
}
\value{
An annotation object
}
\description{
Create a trend line annotation
}
