% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{create_core_story_narrative}
\alias{create_core_story_narrative}
\title{Create core story narrative}
\usage{
create_core_story_narrative(vars, data)
}
\description{
Create core story narrative
}
\keyword{internal}
