% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide.R
\name{describe_mapping_narrative}
\alias{describe_mapping_narrative}
\title{Describe mapping for narrative}
\usage{
describe_mapping_narrative(mapping)
}
\description{
Describe mapping for narrative
}
\keyword{internal}
