% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narrative.R
\name{generate_data_insights}
\alias{generate_data_insights}
\title{Generate data insights for narrative}
\usage{
generate_data_insights(layr_obj)
}
\description{
Generate data insights for narrative
}
\keyword{internal}
