% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide.R
\name{guide_bar}
\alias{guide_bar}
\title{Create a bar guide}
\usage{
guide_bar(mapping = NULL, ...)
}
\arguments{
\item{mapping}{Aesthetic mapping}

\item{...}{Additional parameters}
}
\value{
A guide object
}
\description{
Create a bar guide
}
