% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides.R
\name{guide_density}
\alias{guide_density}
\title{Create a density guide}
\usage{
guide_density(mapping = NULL, ...)
}
\arguments{
\item{mapping}{Aesthetic mapping}

\item{...}{Additional parameters}
}
\value{
A guide object
}
\description{
Create a density guide
}
