% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide.R
\name{guide_line}
\alias{guide_line}
\title{Create a line guide}
\usage{
guide_line(mapping = NULL, ...)
}
\arguments{
\item{mapping}{Aesthetic mapping}

\item{...}{Additional parameters}
}
\value{
A guide object
}
\description{
Create a line guide
}
