% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide.R
\name{guide_scatter}
\alias{guide_scatter}
\title{Create a scatter plot guide}
\usage{
guide_scatter(mapping = NULL, ...)
}
\arguments{
\item{mapping}{Aesthetic mapping}

\item{...}{Additional parameters}
}
\value{
A guide object
}
\description{
Create a scatter plot guide
}
