% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.layr_annotation}
\alias{is.layr_annotation}
\title{Check if object is a layr annotation}
\usage{
is.layr_annotation(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical
}
\description{
Check if object is a layr annotation
}
