% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.layr_guide}
\alias{is.layr_guide}
\title{Check if object is a layr guide}
\usage{
is.layr_guide(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical
}
\description{
Check if object is a layr guide
}
