% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R, R/interactivity.R
\name{make_playable}
\alias{make_playable}
\title{Convert layr visualization to interactive plot}
\usage{
make_playable(layr_obj)

make_playable(layr_obj)
}
\arguments{
\item{layr_obj}{A layr object}
}
\value{
An interactive plotly object

An interactive plotly object
}
\description{
Convert layr visualization to interactive plot

Convert layr visualization to interactive plot
}
\examples{
\dontrun{
if (require(plotly)) {
    viz <- visualize(mtcars, ~mpg ~ wt) |>
        add_guide(guide_scatter())
    make_playable(viz)
}
}
}
