% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narrative.R
\name{narrative}
\alias{narrative}
\title{Generate narrative description of visualization}
\usage{
narrative(layr_obj, preview = FALSE)
}
\arguments{
\item{layr_obj}{A layr object}

\item{preview}{Logical, if TRUE returns short preview}
}
\value{
Character string with narrative
}
\description{
Generate narrative description of visualization
}
