% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides.R
\name{perform_cognitive_check}
\alias{perform_cognitive_check}
\title{Perform cognitive load checks}
\usage{
perform_cognitive_check(layr_obj, guide)
}
\description{
Perform cognitive load checks
}
\keyword{internal}
