% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{view_through}
\alias{view_through}
\title{View through lenses (faceting)}
\usage{
view_through(layr_obj, formula)
}
\arguments{
\item{layr_obj}{A layr object}

\item{formula}{Faceting formula}
}
\value{
Modified layr object
}
\description{
View through lenses (faceting)
}
