% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Create a new layr visualization}
\usage{
visualize(data, formula = NULL, ...)
}
\arguments{
\item{data}{A data frame}

\item{formula}{A formula defining the data story (y ~ x)}

\item{...}{Additional arguments passed to settings}
}
\value{
A layr object
}
\description{
Create a new layr visualization
}
\examples{
\dontrun{
library(gapminder)
viz <- visualize(gapminder, ~lifeExp ~ gdpPercap)
}
}
