% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides.R
\name{add_guide}
\alias{add_guide}
\title{Add a guide to a layr visualization}
\usage{
add_guide(layr_obj, guide)
}
\arguments{
\item{layr_obj}{A layr object}

\item{guide}{A guide object}
}
\value{
Modified layr object
}
\description{
Add a guide to a layr visualization
}
\examples{
\dontrun{
viz <- visualize(mtcars, ~mpg ~ wt) |>
    add_guide(guide_scatter())
}
}
