% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{as_ggplot}
\alias{as_ggplot}
\title{Convert layr object to ggplot2 object}
\usage{
as_ggplot(layr_obj)
}
\arguments{
\item{layr_obj}{A layr object}
}
\value{
A ggplot object
}
\description{
Convert layr object to ggplot2 object
}
\examples{
\dontrun{
viz <- visualize(mtcars, ~mpg ~ wt) |>
    add_guide(guide_scatter())
gg_plot <- as_ggplot(viz)
}
}
