% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactivity.R
\name{make_playable}
\alias{make_playable}
\title{Convert layr visualization to interactive plot}
\usage{
make_playable(layr_obj)
}
\arguments{
\item{layr_obj}{A layr object}
}
\value{
An interactive plotly object
}
\description{
Convert layr visualization to interactive plot
}
